nettype enums

+++ mars/comm/objc/platform_comm.mm
@@ -184,26 +184,26 @@ int getNetInfo() {
 int getNetTypeForStatistics(){
     int type = getNetInfo();
     if (kWifi == type){
-        return NetTypeForStatistics::NETTYPE_WIFI;
+        return (int)NetTypeForStatistics::NETTYPE_WIFI;
     }
     if (kNoNet == type){
-        return NetTypeForStatistics::NETTYPE_NON;
+        return (int)NetTypeForStatistics::NETTYPE_NON;
     }
     
     RadioAccessNetworkInfo rani;
     if (!getCurRadioAccessNetworkInfo(rani)){
-        return NetTypeForStatistics::NETTYPE_NON;
+        return (int)NetTypeForStatistics::NETTYPE_NON;
     }
     
     if (rani.Is2G()){
-        return NetTypeForStatistics::NETTYPE_2G;
+        return (int)NetTypeForStatistics::NETTYPE_2G;
     }else if(rani.Is3G()){
-        return NetTypeForStatistics::NETTYPE_3G;
+        return (int)NetTypeForStatistics::NETTYPE_3G;
     }else if(rani.Is4G()){
-        return NetTypeForStatistics::NETTYPE_4G;
+        return (int)NetTypeForStatistics::NETTYPE_4G;
     }
     
-    return NetTypeForStatistics::NETTYPE_NON;
+    return (int)NetTypeForStatistics::NETTYPE_NON;
 }
 
 unsigned int getSignal(bool isWifi){
